Refactor platform deps(#60338)
