+++ pkg/kubelet/kubelet.go
@@ -143,10 +143,6 @@ const (
 	// ensure this is kept in sync with internal cadvisor housekeeping.
 	evictionMonitoringPeriod = time.Second * 10
 
-	// The path in containers' filesystems where the hosts file is mounted.
-	linuxEtcHostsPath   = "/etc/hosts"
-	windowsEtcHostsPath = "C:\\Windows\\System32\\drivers\\etc\\hosts"
-
 	// Capacity of the channel for receiving pod lifecycle events. This number
 	// is a bit arbitrary and may be adjusted in the future.
 	plegChannelCapacity = 1000
@@ -179,13 +175,6 @@ const (
 )
 
 var etcHostsPath = getContainerEtcHostsPath()
-
-func getContainerEtcHostsPath() string {
-	if sysruntime.GOOS == "windows" {
-		return windowsEtcHostsPath
-	}
-	return linuxEtcHostsPath
-}
 
 // SyncHandler is an interface implemented by Kubelet, for testability
 type SyncHandler interface {
